Practically rewriting the phpbb_posting() function.  It was more complex than it should be, had a number bugs (especially for editing), had a few extra things that were useless (forced in submit_post), and didn't do all that it should.
Edit posts update the forums in our hook for .com.  Will not work on existing posts. #62106

A few other bug fixes.

+++ titania/includes/objects/revision.php
@@ -250,7 +250,7 @@ class titania_revision extends titania_database_object
 							'post_text'				=> $body_reply
 						);
 						$options_reply = array_merge($options_reply, $options);
-						phpbb_posting('reply', $options_reply, true);
+						phpbb_posting('reply', $options_reply);
 					}
 					else
 					{
@@ -559,7 +559,11 @@ class titania_revision extends titania_database_object
 
 		// Delete the queue item
 		$queue = $this->get_queue();
-		$queue->delete();
+
+		if ($queue !== false)
+		{
+			$queue->delete();
+		}
 
 		// Delete the attachment
 		$attachment = new titania_attachment(TITANIA_CONTRIB);
